指摘事項を修正など
